Fix typing error

+++ category_classification/data_utils.py
@@ -19,7 +19,7 @@ def generate_data_from_df(df: pd.DataFrame,
                           product_name_max_length: int,
                           product_name_preprocessing_config: TextPreprocessingConfig):
     ingredient_matrix = process_ingredients(df.known_ingredient_tags,
-                                            ingredient_to_id)
+                                            ingredient_to_id).astype(np.float32)
     product_name_matrix = process_product_name(df.product_name,
                                                nlp=nlp,
                                                token_to_int=product_name_token_to_int,
